Add pip (#1)
* Adding dependencies and locking them using pip

* minor cleanup

* initial travis push

* fix travis install

* fix travis script

* minor update to lock file

* updated pipfile to contain all dependencies

* added pytest and ability to run pytest via pipenv and travis

* adding main() method

* updated README
