+++ .gitignore
@@ -1,0 +1,3 @@
+.DS_store
+.vscode/*
+__pycache__/*
