+++ .travis.yml
@@ -1,0 +1,5 @@
+language: python
+install:
+- pip install pipenv
+- pipenv install .
+script: pytest tests/
