+++ Pipfile
@@ -1,0 +1,21 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+numpy = "*"
+pydicom = "*"
+medpy = "*"
+scipy = "*"
+pytest = "*"
+scikit-learn = "*"
+umap-learn = "*"
+scikit-image = "*"
+pandas = "*"
+mrqy = {path = "."}
+
+[requires]
+python_version = "3.6"
